Enabled profiling and built the sorted list when adding parameters.
The dependency sorting algorithm can be slow (minutes) for large numbers
of randomly ordered parameters (thousands). However, parameters are
normally added in dependency order, so this change builds the default
ordered list as the parameters are added. The function that updates the
ordering will always start with the current list and then sorts based on
which parameters depend on others. So starting with the order "as added"
is generally better than the random hash order from the collection.

Also, some of the previously disabled profiling code was re-enabled to
help evaluate the performance of the ordered list approach. A test of
adding 1000 parameters took about 1.5 seconds. But because this version
is still using the update propagation method, changing the dependency
graph for even one parameter took longer to update than was reasonable
to wait for. Hopefully this will change with the ordered list.