Updated MDL Hash values of CellBlender Test Suite for Blender 2.77
Blender 2.77 generates different mesh orderings for its primitve
objects than Blender 2.76 and previous versions. Some of the tests in
the test suite use these primitive objects (cube and torus) to make
geoemtry for the test (other tests generate their own geometry).

Here is an example between 2.76 and 2.77 using a Cube:

  2.76:

      >>> faces = C.scene.objects['Cube'].data.polygons
      >>> for f in faces:
      ...     vl = [ v for v in f.vertices ]
      ...     print ( vl )
      ...
      [1, 2, 3]
      [7, 6, 5]
      [4, 5, 1]
      [5, 6, 2]
      [2, 6, 7]
      [0, 3, 7]
      [0, 1, 3]
      [4, 7, 5]
      [0, 4, 1]
      [1, 5, 2]
      [3, 2, 7]
      [4, 0, 7]

  2.77:

      >>> faces = C.scene.objects['Cube'].data.polygons
      >>> for f in faces:
      ...     vl = [ v for v in f.vertices ]
      ...     print ( vl )
      ...
      [1, 3, 0]
      [7, 5, 4]
      [4, 1, 0]
      [5, 2, 1]
      [2, 7, 3]
      [0, 7, 4]
      [1, 2, 3]
      [7, 6, 5]
      [4, 5, 1]
      [5, 6, 2]
      [2, 6, 7]
      [0, 3, 7]

The different ordering of the vertices ends up producing different MDL
output and a different MDL hash value. So tests that depend on these
Blender-generated objects have failed in Blender 2.77 while they passed
in Blender 2.76 (and earlier). This commit updates the hash values
stored in the test suite addon to match the values for 2.77. Note that
an alternative "fix" would be to write our own geometry generators for a
cube and a torus. However, the use of Blender's primitive objects is
instructive in the test suite, so that added value favors updating the
hash values rather than using/writing our own mesh generation code. If
Blender continues to change the mesh ordering of their objects in the
future, it might make sense to use our own code.