Added test suite feature to update itself or a copy of itself.
This change provides settings for the test suite to change the expected
hash value into the computed hash value inside a specified file. This is
intended as an easy way to update hash values when the results of a test
change. The file to be changed could be the test suite python file
itself or a copy of that file (for safety). After making the changes,
the two files can be compared to ensure that they are correct.