Added a "clear_ref" function to delete parameter references.
Note that this may have been a problem in previous versions where
parameter references were used. When properties are part of a normal
property group, they are removed automatically when the property is
removed. But with parameter references, the underlying properties must
be removed manually. This has not been done in the past and may account
for some of the bloat that happened as .blend files were used over and
over again.