Removed some failing checks and added a data model test script.
Some of the panel parameter code had been checking to be sure that the
general parameters list wasn't empty. However, this is a legitimate case
when none of the panel parameters depend on any general parameters.

Also added a data model testing script to top of parameter_system.py.
This script can generate some number of parameters that depend on some
number of previous parameters.

This version rechecks everything as it adds each new parameter and is
currently very slow.