Preparation for performance improvement and debugging.
Enabled profiling to measure the time taken in various functions.

Removed "nice" names for some items so they would appear by their real
name in the outliner. The "nice" names are helpful when an item is shown
in a panel. But if the item is not shown in a panel, then it shows up
with the "nice" name in the outliner, but must be referenced by its real
name in the code or the python console.