Added checking to not change properties if not needed.
When a property calculation ends up producing a value that's equivalent
to what's already in the property ... don't change it.

It turns out that setting a property to its current value triggers an
update even though the value hasn't really been changed. Avoiding this
helps speed things up (but it's still very very slow).