Added interpretation of first BNG comment line as containing names.
BioNetGen produces an output file format very similar to MCell plot files.
The BioNetGen format includes a first line that begins with a # character
and then includes the names of the columns afterward. This commit looks
for that # character. If the # is found, then the name of each plot is
taken from that line instead of from the file name and path.