Added simple visibility control to model objects.
The interplay between object selection in the 3D view, object selection
in the model object list, and selection of visibility can be a bit
confusing at the present time. It would be good to work this out at some
point, but it's not clear what the best approach would be. For now, this
commit simply provides a means to change the visibility of the object
listed in the model object selection list.