Added color control to model objects list.
As mentioned in the previous commit, there is some confusion as to which
object should be affected by controls in the model objects list. This is
particularly noticable when operators are placed in a template list.

Operators are essentially global and they don't have any obvious way to
tell which list item they were called from. Our current "add material"
operator adds the material to the item selected in the template list.
But this need not be the item containing the operator that is activated.

To avoid this confusion, the "add material" operator is commented out in
this version. New materials can be added from the display options which
is drawn for the currently selected list item. Once the new material has
been created, then that material is drawn in the proper row and operates
as expected.

The add material operator might work on the currently selected 3D object
(rather than the currently selected list object), but that might be even
more confusing. It would be best if there was a way to keep the selected
3D object and the selected list object to be the same. That hasn't been
done in this commit.