Updated test suite for changes to molecule order and output formatting.
Choosing "Sphere_1" as the default molecule shape changed the order that
the Molecule Glyph test chose the molecules to define and release in the
MDL (it uses the internal molecule glyph enum to define the test). So
this would be an expected change to the MDL. This was confirmed by
examining the MDL to see that the molecule and release sections were
identical but in different places.

Similarly, the MinD/MinE test failed because the previous versions used
a different output formatting command which resulted in the iterations
being printed as 6000000. The newer version produces 6e+06. These are
numerically identical even though their text produces different hashes.

The hash codes for these two tests have been updated to reflect these
changes in this commit.