Only show the surface addition panel if active is a model object
The recently combined object panel and object surface regions panel ends
up showing two sets of add/remove buttons (the top set to add and remove
model objects, and the bottom set to add and remove surface regions for
that object). This could be confusing, and it allowed surface regions to
be defined on non-model objects.

This commit only draws the add/remove portion of the object surface
regions panel if the active object is already a model object (in other
words, object.mcell.include is True).