Removal of Object Name field and cosmetic changes.
There really isn't a good way to rename objects within the CellBlender
interface at this time. This is particularly true since the list of
CellBlender model objects is created on the fly from the "include" flag
in the "mcell" property group on each object. This on-the-fly list is
created by the "model_objects_update" function.

We could map the active object's name directly as a property to be
edited, but we would need to add a callback to it when it's changed so
we can refresh the model objects list. However, since that's a Blender
property, we cannot attach an update function to it (at least I don't
know how to do that). We could create our own "active_object_name"
property, but then that name won't update when objects are selected in
the 3D view.

Renaming of objects needs a little more thought ...