Included updates from sorted_parameters branch (material controls)
Replaced the current cellblender_objects.py file with the version from
the sorted_parameters branch.

This version passes the test suite, however, some of the materials in
the test suite are not shown properly. For example, transparency is
missing from many of the objects. This may be due to the way that the
test suite is creating the materials or it may be due to a problem with
this commit (or recent commits dealing with materials).