Modified test suite to change draw type after adding objects.
Previous changes in CellBlender (mentioned in recent commits) set
different defaults for model objects when they are added to a model.
This caused some of the test suite tests to display differently than
they had without those defaults.

This version of the test suite sets the draw type for objects after
adding them to get the desired draw type.

This version of the test suite also adds a "select_none" function to
deselect all objects in the scene. This keeps the last selected object
from being displayed as orange in the 3D view.