Added an operator to add "count mol in world" for all molecules (mol).
This operator simply goes through the list of molecules and adds a count
statement for each one with the count region being the world. No checks
are made for duplicate count statements at this time.

Because count statements may be somewhat complex, it's not clear what
similar "helper" operators might also be useful. Time will tell.