Consolidated Debug Controls and Deleted lots of commented code.
This version generates about 100 parameters per second in the low
hundreds. However, it shows its non-linear nature around 500. For
example, 600 parameters takes 8 seconds, 1000 takes 25 seconds, and
1500 parameters takes about 72 seconds.

This is much better than the earlier sorted parameter versions, but
there's still lots of room for improvement.

Note that this version is not well tested and there may be many
problems that haven't been discovered yet. Lots of testing is needed.