Call update_panel_expression directly rather than via update in loop
The update_panel_expression function was being called indirectly in the
evaluate_all_pp_expressions function. This indirect call was invoked by
simply assigning ppl[k].expr to itself. This is what happens when the
user changes a given parameter expression field. However, this also
triggers a re-evaluation of all general parameters needed to build a
"globals" dictionary for evaluation of the python version of the
expression. The recreation of the "globals" dictionary isn't needed for
each parameter when updating all panel parameters because it doesn't
change for each panel parameter. So it would be efficient to just create
the globals dictionary once and then use it when evaluating each of the
panel parameters. This commit doesn't do that, but it does provide a
mechanism for doing this by adding an optional "gl" parameter and by
calling the update function directly (rather than indirectly through the
update function). This appears to work which should allow the single
computation of the globals dictionary when re-evaluating all.