Isolated the dictionary building as described in previous commit.
The build_eval_dict function was added to create the evaluation
dictionary once for all general parameters and panel parameters.

The update performance of this version is very good, but creation is
still very slow when creating thousands of parameters.

There may be places where pure Python lists can be used during
calculation and then put back into ID properties when done. The updating
of the dependency order list might be a good place to do this.

This version passes the test suite.