Temporary fix to avoid duplicate parameter names.
This version appends a suffix number ("_2", "_3", ...) to any parameter
name changes that would otherwise match an existing parameter name. So,
for example, if there is an existing parameter named "A", then changing
the name of any other parameter to "A" will immediately produce the name
"A_2" rather than "A".

This is temporary to defer the requirement to maintain the error states
that would otherwise be needed. These error states would need to be
maintained in both Blender's properties and in the data model and all of
its export formats (JSON, for example). The code implementing this
change is marked with a "TODO" comment.