Use bundled Python with matplotlib plotters.
The matplotlib based plotters will now try to use the bundled version of
Python instead of the system version. This will require the user to use
the CellBlender bundle (https://github.com/mcellteam/bundle_cellblender)
instead of the standalone addon. If users want, they can still use the
system version of Python by selecting it in "Settings & Preferences".

Side note: for the sake of having a consistent experience, we should
make sure that we are using the get_python_path function whenever we are
trying to use Python as a subprocess (plotters, sim runners, etc).