Surface Region Debugging and Potential Fix
When rebuilding surface regions from the data model, Blender would
sometimes crash with a segment fault. This would happen in the "find"
call within the add_region_by_name function. It's not clear what causes
this crash, but eliminating the "find" and the subsequent setting of the
active_region_index stops the crash. It's not clear if there are other
side effects from not setting the active_region_index in this context.

This version did pass the test suite, but this change should be
reviewed.