Fixed add_region_by_name to avoic segment fault.
The previous version had retained a reference to a region which had been
invalidated by the "add" method of Blender's CollectionProperty. This
fix saves the name of the region and re-looks it up after the add to
avoid this problem.

Note that this version hasn't had the debugging code "cleaned up" yet.