Speed up parameter parsing by doing some looping rather than recursion.
Python's parser.expr() function returns a deeply nested parse tree. The
CellBlender parameter_system had been using recursion to traverse that
tree and generate the parsed expression that it uses. However, it
appears that most of the layers in this nested tree are not of interest
for expression parsing.

This change replaces some of that recursive searching with a "while"
loop that peels away the layers that are simply 2-tuples containing yet
another tuple inside.

This commit has passed the test suite.