Added some commented code to output parameters with Blender Precision
Because Blender FloatProperties are stored as single precision, they
carry less precision than Blender ID properties (which appear to be
stored as double precision). Now that we are implementing the parameter
system as ID properties, this gives them greater precision than previous
versions. This is not generally a problem, but it does make it difficult
to compare the results of the test suite with previous versions of
CellBlender. This commit adds a small amount of code that can convert
the new ID property versions of parameters using "Blender Precision" by
simply passing them in and out of an RNA FloatProperty dedicated for
that purpose. This is not something that we will want to do, in general,
but it is helpful for testing the new ID Parameter implementation.