Bugfix: Shade Smooth/Flat acting on other objects.
The Shade Smooth/Flat buttons would act on other objects selected
instead of just acting on the molecule selected in the list. For
instance, if I had molecule "vm" highlight in the molecule list and had
a Cube and a Icosphere selected, then all of those objects would be
shaded smooth when I hit the Shade Smooth button.

Side note: we should probably rename these, as they use the same exact
names as the standard Blender operators, which can be really confusing
if you use the space key to search for the shading operators.