Fixing test suite to work with new parameter system
The previous vertion tested whether the parameter system was using RNA
properties or ID properties with this:

  if 'active_elist' in ps

Where the active_elist was a new member data item added to the ID
property version. However, bcause Blender doesn't really add RNA
properties to a property group until they are changed, this would
false even when the ID property version was in use. So checking
whether a name is "in" the object's dictionary is not reliable.

Instead, it appears that using the "dir" function on the object,
does return the proper result even if the RNA property hasn't been
changed from its default. So the following test appears to work:

  if 'active_elist' in dir(ps)

This change appears to fix the test suite problem (although further
testing is needed).