Modified test suite to handle precision differences
Due to the different representations of ID and RNA parameters in
Blender, the results of repeated parameter operations can produce
different parameter values even when they are being similarly truncated
when exported to MDL. This currently only happens in one of the test
cases within the test suite. This commit adds a small check in the test
suite code to expect different results for that particular test. All
other results are identical (assuming parameters are exported in
Blender's RNA property representation).

Details:

The "Non-Geometry Tests" section contains a test (number 8) called "100
Pars Using 3 Each". This test generates a series of parameters that each
depend on the 3 previous parameters (typically a performance test). The
first few parameters look like this:

  a = 1e-6
  b = 1e-6+a
  c = 1e-6+a+b
  d = 1e-6+a+b+c
  e = 1e-6+b+c+d
  f = 1e-6+c+d+e
  : :    :

They continue on and eventually begin using names like "P_42" (after the
alphabet is exceeded). Both the RNA and ID implementations produce
identical results for parameters a through g. But they begin to differ
at paramter "h" as shown in these two MDL segments:

  Development Branch:
    /* DEFINE PARAMETERS */
    a = 9.99999997475243e-07    /* Parameter a    units=u */
    b = 1.99999999495049e-06    /* Parameter b    units=u */
    c = 3.99999998990097e-06    /* Parameter c    units=u */
    d = 7.99999997980194e-06    /* Parameter d    units=u */
    e = 1.49999996210681e-05    /* Parameter e    units=u */
    f = 2.80000003840541e-05    /* Parameter f    units=u */
    g = 5.19999994139653e-05    /* Parameter g    units=u */
    h = 9.59999961196445e-05    /* Parameter h    units=u */
    i = 0.00017699999443721    /* Parameter i    units=u */
    j = 0.000325999979395419    /* Parameter j    units=u */
    k = 0.00059999997029081    /* Parameter k    units=u */
      :
    P_97 = 6.35358247688035e+19    /* Parameter P_97    units=u */
    P_98 = 1.16860599399661e+20    /* Parameter P_98    units=u */
    P_99 = 2.1494015041421e+20    /* Parameter P_99    units=u */

  Sorted Parameters Branch:

    /* DEFINE PARAMETERS */
    a = 9.99999997475243e-07    /* Parameter a    units=u */
    b = 1.99999999495049e-06    /* Parameter b    units=u */
    c = 3.99999998990097e-06    /* Parameter c    units=u */
    d = 7.99999997980194e-06    /* Parameter d    units=u */
    e = 1.49999996210681e-05    /* Parameter e    units=u */
    f = 2.80000003840541e-05    /* Parameter f    units=u */
    g = 5.19999994139653e-05    /* Parameter g    units=u */
    h = 9.60000033956021e-05    /* Parameter h    units=u */
    i = 0.00017699999443721    /* Parameter i    units=u */
    j = 0.00032600000849925    /* Parameter j    units=u */
    k = 0.000600000028498471    /* Parameter k    units=u */
      :
    P_97 = 6.35358071766175e+19    /* Parameter P_97    units=u */
    P_98 = 1.16860573011382e+20    /* Parameter P_98    units=u */
    P_99 = 2.14940097637652e+20    /* Parameter P_99    units=u */

As shown here, they are identical until parameter "h". While these
differences have not been analyzed in detail, they are consistent with
the kinds of differences that might be seen between double and float
representations used in that kind of repeated calculation.

As a result of this examination (and the fact that all other test suite
results are identically equal), this commit includes a test for the
parameter implementation (RNA or ID) and expects different MDL file
hashes between the two versions. In other words, both the development
and the sorted_parameters branches should pass this same test suite.