Reset mol viz file path after "save as" - Preliminary
When the "save as" menu command is used to save a blend file by a
different name, the molecule visualization code continued to read the
"mol viz" data from the location corresponding to the previous file name
until the simulation was run again. This could be confusing. This commit
adds a "save post" handler to compare the mol viz directory path with
the path of the newly saved blend file. If they are consistent (as in a
regular "save"), then nothing is done. But if they are inconsistent (as
in a "save as"), then the mol viz information is erased. Note that this
"erasing" is currently done by setting the global_mol_file_list to an
empty list and seting the mol_file_dir to an empty string. It's not
clear whether this is sufficient to properly reset the state to the
condition of not having been run (which is the ultimate intention).

Note that this action is only taken when the mol viz directory was NOT
manually selected. When a mol viz directory is manually selected, it is
assumed that the user was viewing that data for a specific purpose, and
the code cannot "read the user's mind" in that case.

Also note that this implementation precludes the possibility of saving
as a different file name and then re-saving as the original file name
with the intention of re-enabling the reading of viz data from the
original directory. This is because the code has no way of knowing
whether the new "old" directory really matches the contents of the
model. There is currently no hashing of the model and matching with the
hashing of the results to unambiguously tie them together. It is just
assumed that the two correspond when a model has been run. Therefore, if
it is desired to re-view the original data after a "save as" operation,
the best approach is to simply re-open the original .blend file and
refresh the molecules.

Also note that this current implementation does not remove the molecule
position mesh from the current scene. These are part of Blender's data,
and are normally included in the "save as" operation. They will not be
refreshable in that case (since the mol file list is empty), and that
could be considered as an indication that the model hasn't been run
under the new name. More properly, these could be removed to give a
much clearer indication that the model has not been run. There may be
other ways to convey this, so it hasn't been done until a decision is
made regarding the desired behavior.

Finally, as noted in the original paragraph, it's not clear what the new
state should be. There appear to be a number of different Blender
property structures in the mol viz code that may not be in use any more.
They may have been replaced with the global_mol_file_list system for
performance purposes and not removed (in case the performance problem is
resolved). So the entire module could use some careful scrutiny.