Added functions to save and restore visibility, selection, and active
Some of the operations performed in CellBlender affect the visibility,
selection, and active object(s). This commit adds a "get_object_status"
function which returns a dictionary of object/status for these items. It
also adds a "restore_object_status" function to restore the values from
such a dictionary.

At this point, these functions are only used in cellblender_objects
itself, and it does not preserve this information across data model
upgrades. It could possibly be called around the upgrades to preserve
information as found in the .blend file (not needed separate storage in
the data model).