Variable change needed for sorted parameters branch.
Since this code is shared with the sorted_parameters branch, it could be
changed in either location and then merged into the other. However,
since the sorted_parameters branch contains newer code than development,
the merge should always be from development into sorted parameters
(until the final merge which brings the two branches together). So any
changes that may be needed for the sorted_parameters branch should be
made to development first and then merged into sorted_paramters. That's
the case here with changes that do not affect the development branch,
but are needed by the sorted_parameters branch.