Make molecule viz objects respect layer settings.
Previously, the molecule visualization objects (glyph and positions)
would appear on every layer that was selected. You could move them
temporarily, but they would move back as soon as the frame changed (as
happens with an animation). This made it impossible to use render layers
effectively for compositing. In addition to advanced rendering
techniques, layers are an incredibly useful tool for creating logical
groups when you are dealing with complex models that have dozens of
different meshes and molecules.