Added a check to ensure that test hasn't been run before saving startup.
One of the common mistakes when running the test suite is accidentally
saving the startup file after a test has been run. This saved startup
file is then tainted with results of the previous test, and may not give
proper results when used as the starting point for a subsequent test.

This commit checks that the test status is "unknown" ("?") before saving
the startup file via the handy "Save Startup" button. This helps prevent
overwriting the startup file with the results of a previous test (failed
or not). If it's really desired to save the startup file from that
state, it can be easily accomplished through the normal File menu
command (File / Save Startup File).