Changed order of parameters saved in data model
Previous versions had saved parameters in the data model in sorted order
so they would follow the "define before use" pattern. However, this ends
up reordering the parameters when a .blend file is upgraded (because the
original user order had been lost).

This commit renames the old "build_data_model_from_properties" function
to "build_ordered_data_model_from_properties" and creates a duplicate of
the original function which stores the parameters in user order.

This may not be the best solution since it leaves the data model
parameters in a state where parameters are not guaranteed to be defined
before used. But it does restore the parameters in the order that the
user had defined them (which is also a valuable property). This is not a
problem for CellBlender but might be inconvenient for other programs
that expect the "define before use" pattern.

Maybe a better solution (requiring more storage) would export an
additional list that either specifies the user order (if the parameters
are exported in dependency order) or the dependency order (if the
parameters are stored in user order). The current internal
representation is the latter (parameters are listed and displayed in
user order, but a separate list maintains the dependency order).