Added dependency ordered ID list to data model in "_extras"
The "_extras" key has been used to store data that might be convenient
for external programs using the data model. In this case, the dependency
ordered ID list is helpful for determining the order that parameters
should be defined such that all dependencies are satisfied as each new
parameter is added. CellBlender already builds this list internally for
its own use. Previous versions had exported the parameter list in that
dependency order as a convenience for external programs. However, this
had the unfortunate effect of losing the user's defined order when
upgrading from a data model.

More recent versions now export the data model retaining the user's
original order. This lost the dependency order that had been computed
and maintained inside CellBlender. The current commit adds that
depencency order in a separate list in the "_extras" dictionary of the
entire parameter system. CellBlender itself does not (currently) make
use of that dictionary and it recomputes its dependency list internally.
So this new list is primarily for use by external code.