Fixed viz_data_save_post handler to not clear viz on save
Commit 707cc7a (Jul 12, 2016) added a viz_data_save_post function intended
to empty the visualization data when a file was saved with a different
file name. That function was intended to detect a difference between the
shared portions of the .blend file path and the viz output file path. If
the shared portions were different, then it emptied the viz data forcing
a reload to see the viz data again. That detection was flawed, and a
reload was generally required even when the file paths were the same.
This commit more properly detects the path differences and does not
require the reloading of viz data when saving to the same file name.