Began supporting dynamic geometry directly in CellBlender
This commit adds code to read geometry for a single object from either
Blender's own mesh data or by calling a user's text script (if one is
provided).

Each object has a "dynamic" flag, and a script name. When the dynamic
flag is set, CellBlender will check the script name. If the name is
blank, then CellBlender will attempt to get the variable geometry for
that object from the Blender object. If the script name is not blank,
then CellBlender will call the script to get the geometry.

In either case, CellBlender will generate the MDL files for all of the
iterations requested in the run when it exports the other MDL. If any
"dynamic" flags are set, then CellBlender will read the dynamic geometry
from the associated MDL files to display during replay.

This is a very preliminary implementation based on box_test.py found in
the test_suite folder. There are some known problems that haven't been
worked out yet, but since it only functions when the "dynamic" flag is
set for an object, it should run fine if that flag is simply not set
(which is the default).