Added some comments describing the developing dynamic geometry API.
Comments were added to explain the role of the user's dynamic geometry
code, and the previous variable "i" was replaced with "frame_number" to
better describe its purpose.

There are still a lot of unresolved issues with dynamic geometry, but
this code lets people build dynamic geometry models fairly easily
(assuming that they have the code for the geometry itself).