Eliminated setting of active_obj_index in build properties loop.
The previous version set the active object index to the end of the list
after each new MCellModelObjectsProperty was added to the object_list.
This triggered the "active_obj_index_changed" callback which is
currently set to unhide the object, select it, and make it active. This
is appropriate for selecting objects from the list manually, but adds a
lot of overhead when adding them in a loop. It might also be forcing a
redraw of the panel each time as well.

This commit moves the updating of the active_obj_index to the end of the
loop. Additionally, it checks to see if the active_obj_index is within
the newly (re)created list of objects. If so, then it doesn't change it.
If not, then it sets it to the last object added.

This commit dramatically speeds up the adding of objects which prints:

  "Data model contains " + m["name"]

That loop took a long time in the previous version, but is much faster
with this change.