Added some "check before set" tests to minimize property changes
Previous code simply set values for hide and select as appropriate
without checking whether they might already be hidden and selected.
It's not clear which is more efficient, but experience has shown that
changing a flag might be more expensive than checking it. So this
version checks those flags and only sets them as needed.