Used self in "check_model_object" (in cellblender_objects.py)
This is a common idiom within CellBlender:

  mcell = context.scene.mcell
  model_object_list = mcell.model_objects.object_list
  model_object = model_object_list[mcell.model_objects.active_obj_index]

But this requires that the active_obj_index always point to the
currently "active" object. This is fine with Blender's user interface,
but it can be a problem when used to loop through all model objects.
This is because changing the active_obj_index will call any associated
callback functions which can cause considerable overhead.

An earlier commit removed setting of the active_obj_index during tight
loops to improve performance, but since the "check_model_object"
function was relying on that idiom, it was failing (error messages). It
turns out that the "self" passed to "check_model_object" was already an
MCellModelObjectsProperty, so there was no need to get it from the
active_obj_index via context.scene.mcell. It could be more easily (and
reliably) gotten from the "self" passed as the first parameter. This
commit implements that change.

Note that this might be something to watch for throughout CellBlender.