Stress testing and more efficient model object updating.
This commit adds a first cut "stress test" for model objects in
CellBlender. It also adds a more efficient model_objects_update
function. However, this new function doesn't end up sorting the
model objects as the old one did.

The model_objects_update function is tasked with searching through all
objects in a scene and finding those containing an mcell property group.

Early versions just rebuilt the list from scratch. More recent versions
have had to save additional information before rebuilding. Finally, this
version  implements an add/subtract strategy to adjust the existing list
based on set differences. However, this gives up the easy pure-python
sorting present in the previous version. Furthermore, it appears that
the typical usage of this routine (in the test suite) is after adding
each new object. So the greater performance gain would be delaying the
call to this function altogether until the final add has been done.

So for all these reasons, it might be better to revert to the previous
version of model_object_update. This version is being saved to document
the strategy for possible future incorporation.