Reverted model_objects_update to previous version.
The stress test that suggested efficiency problems in
model_objects_update might be itself flawed. The CellBlender test suite
is designed to test setting of properties and invocation of operators.
These are done manually, and testing them in tight loops for efficency
isn't the best way to find speed problems.